Added Castable interface. Added casting/activating.
The castable interface is used to control spells/skills that can be
cast/activated. Also added a function in Functions that can cast spells
and skills that are currently chosen on the hotbar. Added support for
spells and skills with duration. The player instance now controls the
cooldowns and active "buffs".